# openapi-specification.yaml
openapi: 3.1.0
info:
  title: ProcGrid B2B E-commerce Platform API
  description: |
    Production-grade B2B e-commerce platform for producer-buyer agricultural product trading.
    
    ## Authentication
    All protected endpoints require JWT Bearer token obtained from Keycloak authentication server.
    
    ## Rate Limiting
    - Public endpoints: 100 requests/minute
    - Authenticated endpoints: 1000 requests/minute
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details for HTTP APIs standard.
  version: 1.0.0
  contact:
    name: ProcGrid API Support
    email: api-support@procgrid.com
    url: https://procgrid.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://procgrid.com/terms
servers:
  - url: https://api.procgrid.com/v1
    description: Production server
  - url: https://staging-api.procgrid.com/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /register/producer:
    post:
      tags:
        - User Registration
      summary: Register producer/farmer
      description: Register new producer with farm details
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerRegistrationRequest'
      responses:
        '201':
          description: Producer registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /register/buyer:
    post:
      tags:
        - User Registration
      summary: Register buyer/merchant
      description: Register new buyer with business details
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyerRegistrationRequest'
      responses:
        '201':
          description: Buyer registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /users/{id}:
    get:
      tags:
        - User Management
      summary: Get user profile
      description: Retrieve user profile details
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - User Management
      summary: Update user profile
      description: Update user profile information
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Keycloak authentication
  parameters:
    UserIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: "user_12345"
...
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message
        errorCode:
          type: string
          description: Error code for failed requests
          nullable: true
        data:
          description: Response data
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      type: object
      required:
        - success
        - message
        - errorCode
        - timestamp
        - path
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error description
        errorCode:
          type: string
          description: Specific error code
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: API endpoint path

    # ============ AUTHENTICATION SCHEMAS ============
...
    UserRole:
      type: string
      enum:
        - PRODUCER
        - BUYER
        - ADMIN
      description: User role in the system
    UserStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - SUSPENDED
        - PENDING_VERIFICATION
      description: User account status
    KycStatus:
      type: string
      enum:
        - NOT_STARTED
        - SUBMITTED
        - UNDER_REVIEW
        - VERIFIED
        - REJECTED
      description: KYC verification status
...
    ProducerRegistrationRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - phoneNumber
        - farmName
        - location
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 6
          maxLength: 128
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        phoneNumber:
          type: string
          pattern: '^\\+?[1-9]\\d{1,14}$'
        farmName:
          type: string
          maxLength: 200
        location:
          type: string
          maxLength: 500
        gstNumber:
          type: string
          pattern: '^[0-9]{2}[A-Z]{5}[0-9]{4}[A-Z]{1}[1-9A-Z]{1}Z[0-9A-Z]{1}$'
          nullable: true
    BuyerRegistrationRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - phoneNumber
        - businessName
        - businessType
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 6
          maxLength: 128
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        phoneNumber:
          type: string
          pattern: '^\\+?[1-9]\\d{1,14}$'
        businessName:
          type: string
          maxLength: 200
        businessType:
          type: string
          enum: [WHOLESALER, RETAILER, PROCESSOR, EXPORTER]
        gstNumber:
          type: string
          pattern: '^[0-9]{2}[A-Z]{5}[0-9]{4}[A-Z]{1}[1-9A-Z]{1}Z[0-9A-Z]{1}$'
          nullable: true
...
    UserProfile:
      type: object
      required:
        - userId
        - email
        - firstName
        - lastName
        - phoneNumber
        - role
        - status
        - kycStatus
        - createdAt
      properties:
        userId:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        kycStatus:
          $ref: '#/components/schemas/KycStatus'
        profileImage:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserUpdateRequest:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        phoneNumber:
          type: string
          pattern: '^\\+?[1-9]\\d{1,14}$'
        profileImage:
          type: string
          format: uri
...
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Resource not found"
            errorCode: "USER_001"
            timestamp: "2025-11-15T10:30:00Z"
            path: "/api/v1/users/999"
    ConflictError:
      description: Conflict - Resource state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Resource already exists"
            errorCode: "USER_002"
            timestamp: "2025-11-15T10:30:00Z"
            path: "/api/v1/register/producer"
    ValidationError:
      description: Validation error - Invalid input data
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  data:
                    type: object
                    additionalProperties:
                      type: string
          example:
            success: false
            message: "Validation failed"
            errorCode: "VAL_001"
            data:
              email: "Invalid email format"
              password: "Password must be at least 6 characters"
            timestamp: "2025-11-15T10:30:00Z"
            path: "/api/v1/register/buyer"
...